Tax spaces and free parking
I added functions to deduct and add money to the player, and what
happens when you land on the tax spaces and started a big else if
structure for spaces like it. Once the messages label is working
properly we should be able to just uncomment that line where the text
is set. I also coded Free parking because it was super easy and needed
to be done.